+++ SmartScraper/src/component/CertificateFactory/CertificateGenerator.js
@@ -1,13 +1,15 @@
 const fs = require("fs");
+const path = require("path");
 const {
   CertificateGeneratorFactory,
-} = require("./CertificateGeneratorFactory");
+} = require("./CertificateGeneratorFactory.js");
 
 class CertificateGenerator {
   constructor(options, hostname, certDir) {
     this.options = options;
     this.hostname = hostname;
     this.certsDir = certDir;
+    this.run = this.run.bind(this);
   }
 
   run() {
@@ -20,22 +22,24 @@ class CertificateGenerator {
     }
     // Get correct generator form the CertificateGeneratorFactory
     // get certificate generator instance
-    const certificateGenerator =
-      CertificateGeneratorFactory.getCertificateGenerator(
-        this.options,
-        this.hostname,
-        this.certsDir
-      );
+    
+    const certificateGeneratorFactory = new CertificateGeneratorFactory();
+    const certificateGenerator = certificateGeneratorFactory.getCertificateGenerator(
+      this.options,
+      this.hostname,
+      this.certsDir
+    );
     // use certificateGenerator instance to generate the certificate
     return certificateGenerator.generateCertificate();
   }
+
 }
 
 module.exports = { CertificateGenerator };
-
-const certGen = new CertificateGenerator(
-  "-f",
-  "./certs/example.req",
-  "./certs"
-);
-certGen.run();
+//
+// const certGen = new CertificateGenerator(
+//   "-f",
+//   "./certs/example.req",
+//   "./certs"
+// );
+// certGen.run();