+++ SmartScraper/src/component/CertificateFactory/CertificateGeneratorFactory.js
@@ -1,13 +1,13 @@
-const { StringCertificateGenerator } = require("./StringCertificateGenerator");
-const { FileCertificateGenerator } = require("./FileCertificateGenerator");
+const { StringCertificateGenerator } = require("./StringCertificateGenerator.js");
+const { FileCertificateGenerator } = require("./FileCertificateGenerator.js");
 const {
   DefaultCertificateGenerator,
-} = require("./DefaultCertificateGenerator");
-const { ConfigCertificateGenerator } = require("./ConfigCertificateGenerator");
+} = require("./DefaultCertificateGenerator.js");
+const { ConfigCertificateGenerator } = require("./ConfigCertificateGenerator.js");
 
 // Define Certificate Generator Strategy
 class CertificateGeneratorFactory {
-  static getCertificateGenerator(options, hostname, certDir) {
+  getCertificateGenerator(options, hostname, certDir) {
     switch (options) {
     case "-f" || "--file": // passing a file containing the hostname.
       return new FileCertificateGenerator(options, hostname, certDir);
@@ -15,8 +15,9 @@ class CertificateGeneratorFactory {
       return new StringCertificateGenerator(options, hostname, certDir);
     case "-d" || "--default": // default hostname used is example.com, no second parameter should be entered
       // eslint-disable-next-line no-case-declarations
-      return new DefaultCertificateGenerator(options, "example.com", certDir);
-    case "-c" || "--config": // config file to be used in certificate generation. If no config file is provided one will be created from hostname.
+      return new DefaultCertificateGenerator(options, hostname, certDir);
+    case "-c" || "--config": // hostname should contain the config file path to be used in certificate generation.
+      // If no config file path is found "example.com" will be used to create certificates.
       // eslint-disable-next-line no-case-declarations
       return new ConfigCertificateGenerator(options, hostname, certDir);
     default:
@@ -32,7 +33,9 @@ class CertificateGeneratorFactory {
         "-d | --default: default hostname used is example.com, no second parameter should be entered."
       );
       console.log(
-        "-c | --config: config file to be used in certificate generation. If no config file is provided one will be created from hostname."
+        "-c | --config: config file to be used in certificate generation.\n" +
+          "If no config file is provided one will be created from hostname.\n" +
+          "The hostname parameter should be the path to the config file."
       );
       console.log("-h | --help: this help message.");
       process.exit(1);