+++ SmartScraper/src/component/ChatBotDialog.js
@@ -1,5 +1,5 @@
 import React, { Component, useState } from "react";
-import { Responsive } from "./Responsive";
+import { Responsive } from "./Responsive.js";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import Modal from "react-bootstrap/Modal";