+++ SmartScraper/src/component/Strategy/StreamStrategy.ts
@@ -7,16 +7,13 @@ import {IStrategy} from "./IStrategy.ts";
  * This class implements a 64 bit byte array (Uint8Array) of the Strategy pattern hierarchy
  * and should be implemented or extended into all subsequent classes.
  */
-interface StreamStrategy extends IStrategy {
-
+export default interface StreamStrategy extends IStrategy {
     /**
      * Send a request to Gpt3 API endpoints
      * @param {{prompt: string, layman: false}} options
      * @param {{element: Element}} streamOptions??
      * @returns  {{Promise<string | JSON | ReadableStream<object> | JSX.Element | JSX.Element[] | HTMLElement | void>}} JSX element(s), empty array, or string.
      */
-    streamRequest(options: { prompt: string; layman: false }, streamOptions?: { element: Element }) :
+    streamRequest(options: { prompt: string; layman: false }, streamOptions?: { element: Element }):
         Promise<string | JSON | ReadableStream<object> | JSX.Element | JSX.Element[] | HTMLElement | void>
 }
-
-export {StreamStrategy};